Moved all the server setup to setup.ts
