+++ src/server.ts
@@ -1,7 +1,45 @@
 import { ApolloServerPluginDrainHttpServer, Config } from 'apollo-server-core';
 import { ApolloServer, ExpressContext } from 'apollo-server-express';
+import { config as makeDotenvAvailable } from 'dotenv';
 import express from 'express';
 import http from 'http';
+import { DataSource } from 'typeorm';
+import { DEFAULT_GRAPHQL_PATH, DEFAULT_SERVER_PORT } from './consts';
+import { appDataSource } from './data-source';
+import { helloResolver } from './resolvers/hello.resolver';
+import { userResolver } from './resolvers/user.resolver';
+import { helloTypeDef } from './types/hello.type';
+import { userTypeDef } from './types/user.type';
+
+export interface GraphqlContext {
+  dataSource: DataSource;
+}
+
+export async function setupServer() {
+  makeDotenvAvailable();
+  const server = await configAndInitilizeServer();
+  return server;
+}
+
+async function configAndInitilizeServer() {
+  const context: GraphqlContext = {
+    dataSource: await appDataSource.initialize(),
+  };
+
+  const PORT = process.env.APP_PORT ?? DEFAULT_SERVER_PORT;
+  const PATH = process.env.GRAPHQL_PATH ?? DEFAULT_GRAPHQL_PATH;
+
+  const server = await startApolloServer(
+    [userTypeDef, helloTypeDef],
+    [userResolver, helloResolver],
+    context,
+    PORT,
+    PATH,
+  );
+
+  console.log(`Graphql server is running on: http://localhost:${PORT}${PATH}`);
+  return server;
+}
 
 export async function startApolloServer(
   typeDefs: Config<ExpressContext>['typeDefs'],
@@ -25,4 +63,5 @@ export async function startApolloServer(
   });
 
   httpServer.listen({ port });
+  return httpServer;
 }
